analytics: Move unnecessary classes from usage publisher to gateway
Usage publisher contained most of the analytics related handlers and
mediators. Which required a dependency to gateway module from usage
publisher module. All of these classes are moved to gateway module
to correct usage of these modules and avoid the cyclic dependency in
Websocket API analytics implementation.
