+++ components/apimgt/org.wso2.carbon.apimgt.gateway/pom.xml
@@ -118,6 +118,11 @@
             <groupId>org.wso2.carbon.metrics</groupId>
             <artifactId>org.wso2.carbon.metrics.manager</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.wso2.carbon.apimgt</groupId>
+            <artifactId>org.wso2.carbon.apimgt.usage.publisher</artifactId>
+        </dependency>
+
     </dependencies>
 
     <build>
