+++ components/apimgt/org.wso2.carbon.apimgt.usage/org.wso2.carbon.apimgt.usage.publisher/src/main/java/org/wso2/carbon/apimgt/usage/publisher/APIMgtUsageDataPublisher.java
@@ -18,7 +18,6 @@
 package org.wso2.carbon.apimgt.usage.publisher;
 
 import org.wso2.carbon.apimgt.api.APIManagementException;
-import org.wso2.carbon.apimgt.gateway.dto.ExecutionTimePublisherDTO;
 import org.wso2.carbon.apimgt.usage.publisher.dto.*;
 
 public interface APIMgtUsageDataPublisher {
