+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/admin/modules/manage-alerts/alerts.jag
@@ -61,7 +61,7 @@ var unSubscribe = function (userName){
             };
 
         }
-        var apiDataPublisher = org.wso2.carbon.apimgt.usage.publisher.AlertTypesPublisher;
+        var apiDataPublisher = org.wso2.carbon.apimgt.gateway.mediators.AlertTypesPublisher;
         var apiDatapublisherInstance = new apiDataPublisher();
 
         apiDatapublisherInstance.unSubscribe(userName,agent);
@@ -97,7 +97,7 @@ var saveAlertsTypes  = function (checkedAlertList, emailList,userName,checkedAle
             };
 
         }
-        var apiDataPublisher = org.wso2.carbon.apimgt.usage.publisher.AlertTypesPublisher;
+        var apiDataPublisher = org.wso2.carbon.apimgt.gateway.mediators.AlertTypesPublisher;
         var apiDatapublisherInstance = new apiDataPublisher();
         //if it is publisher "p" if it is store agent should be "s"
