+++ features/apimgt/org.wso2.carbon.apimgt.store.feature/src/main/resources/store/modules/manage-alerts/alerts.jag
@@ -63,7 +63,7 @@ var unSubscribe = function (userName){
             };
 
         }
-        var apiDataPublisher = org.wso2.carbon.apimgt.usage.publisher.AlertTypesPublisher;
+        var apiDataPublisher = org.wso2.carbon.apimgt.gateway.mediators.AlertTypesPublisher;
         var apiDatapublisherInstance = new apiDataPublisher();
 
         apiDatapublisherInstance.unSubscribe(userName,agent);
@@ -94,7 +94,7 @@ var saveAlertsTypes  = function (checkedAlertList, emailList,userName,checkedAle
         var APIManagerFactory = org.wso2.carbon.apimgt.impl.APIManagerFactory.getInstance();
         var apiProviderImpl = APIManagerFactory.getAPIProvider(tenantAwareUserName);
 
-        var apiDataPublisher = org.wso2.carbon.apimgt.usage.publisher.AlertTypesPublisher;
+        var apiDataPublisher = org.wso2.carbon.apimgt.gateway.mediators.AlertTypesPublisher;
 
         var apiDatapublisherInstance = new apiDataPublisher();
         //if it is publisher "p" if it is store agent should be "s"
